refactor(briefs): search briefs with a plain query function
Replace the search store interface, class, and process registry entry
with one searchConversationBriefs function that takes the database, so
the tool and tests query the table directly.
